Basic IA Added and Enemy Spawner
I added a basic IA to follow the player if it is in sight. If the player is behind a wall the enemy won't move until it has visual.

I also added some sound effects to improve the gamefeel and the player loses health when an enemy bomb hits it.